Create IAM role for cluster with custom assume policy.
This change allows calling modules to restrict the assume policy e.g. to the same account by adding a condition on the sourceAccount.
The module will create an IAM role and pass it to the downstream module (terraform-eks), as the downstream module only accepts a custom role and does not allow the provision of our own policy.